requirement.txt
h5py==2.10.0
ipykernel==6.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.5
matplotlib==3.5.0
matplotlib-inline==0.1.3
numpy==1.18.0
opencv-python==4.4.0.44
pandas @ file:///rapids/pandas-1.2.5-cp38-cp38-manylinux_2_5_x86_64.manylinux1_x86_64.whl
Pillow==9.0.1
pyzmq==22.3.0
regex==2022.1.18
requests==2.27.1
rmm @ file:///rapids/rmm-21.12.0a0%2B31.g0acbd51-cp38-cp38-linux_x86_64.whl
sacremoses==0.0.47
scikit-image==0.17.1
scikit-learn @ file:///rapids/scikit_learn-0.24.0-cp38-cp38-manylinux2010_x86_64.whl
scipy==1.5.0
seaborn==0.12.2
SimpleITK==2.2.1
tensorboard==1.15.0
tqdm==4.62.3
pyradiomics==2.2